Simplify interface
