+++ src/dust_rng.cpp
@@ -2,7 +2,7 @@
 #include <cpp11/external_pointer.hpp>
 #include <cpp11/raws.hpp>
 #include <dust/rng.hpp>
-#include <dust/interface.hpp>
+#include <dust/rng_interface.hpp>
 
 typedef dust::pRNG<double> dust_rng_t;
 typedef cpp11::external_pointer<dust_rng_t> dust_rng_ptr_t;
